+++ framework/packages/staking/src/msg.rs
@@ -89,8 +89,47 @@ pub enum StakingQueryMsg {
 }
 
 #[cosmwasm_schema::cw_serde]
+#[non_exhaustive]
+pub enum StakingTarget {
+    Contract(Addr),
+    Id(u64),
+}
+
+impl StakingTarget {
+    pub fn expect_contract(self) -> abstract_core::AbstractResult<Addr> {
+        match self {
+            StakingTarget::Contract(addr) => Ok(addr),
+            _ => Err(abstract_core::AbstractError::Assert(
+                "Staking target is not a contract address.".into(),
+            )),
+        }
+    }
+
+    pub fn expect_id(self) -> abstract_core::AbstractResult<u64> {
+        match self {
+            StakingTarget::Id(id) => Ok(id),
+            _ => Err(abstract_core::AbstractError::Assert(
+                "Staking target is not an pool ID.".into(),
+            )),
+        }
+    }
+}
+
+impl From<u64> for StakingTarget {
+    fn from(value: u64) -> Self {
+        Self::Id(value)
+    }
+}
+
+impl From<Addr> for StakingTarget {
+    fn from(value: Addr) -> Self {
+        Self::Contract(value)
+    }
+}
+
+#[cosmwasm_schema::cw_serde]
 pub struct StakingInfoResponse {
-    pub staking_contract_address: Addr,
+    pub staking_target: StakingTarget,
     pub staking_token: AssetInfo,
     pub unbonding_periods: Option<Vec<Duration>>,
     pub max_claims: Option<u32>,