+++ integrations/wyndex/packages/wyndex-adapter/src/staking.rs
@@ -2,7 +2,7 @@ use crate::AVAILABLE_CHAINS;
 pub use crate::WYNDEX;
 use abstract_sdk::core::objects::LpToken;
 use abstract_staking_adapter_traits::Identify;
-use cosmwasm_std::{Addr, Env};
+use cosmwasm_std::{Addr, Env, MessageInfo};
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct WynDex {
@@ -67,7 +67,9 @@ impl CwStakingCommand for WynDex {
         &mut self,
         deps: Deps,
         env: Env,
+        _info: Option<MessageInfo>,
         ans_host: &AnsHost,
+        _abstract_registry: Addr,
         lp_token: AssetEntry,
     ) -> std::result::Result<(), AbstractSdkError> {
         self.staking_contract_address = self.staking_contract_address(deps, ans_host, &lp_token)?;
@@ -151,7 +153,7 @@ impl CwStakingCommand for WynDex {
         )?;
 
         Ok(StakingInfoResponse {
-            staking_contract_address: self.staking_contract_address.clone(),
+            staking_target: self.staking_contract_address.clone().into(),
             staking_token: AssetInfo::Cw20(self.lp_token_address.clone()),
             unbonding_periods: Some(
                 bonding_info_resp