+++ framework/CHANGELOG.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+- Updated fetch_data arguments of CwStakingCommand
 
 ### Fixed
+- Partially fixed cw-staking for Osmosis
 
 ## [0.17.2] - 2023-07-27